refactored parsing code into new branch
